+++ src/xenia/cpu/backend/a64/a64_backend.h
@@ -199,7 +199,7 @@ class A64Backend : public Backend {
   void set_trace_data_enabled(bool value) override;
   bool trace_func_enabled() const override;
   void set_trace_func_enabled(bool value) override;
-  std::string FormatSequenceKey(uint32_t key) const override;
+  std::string FormatSequenceKey(uint64_t key) const override;
 
   void RecordMMIOExceptionForGuestInstruction(void* host_address);
