+++ src/xenia/cpu/backend/a64/a64_sequences.h
@@ -47,7 +47,7 @@ bool SelectSequence(A64Emitter* e, const hir::Instr* i,
 
 // Renders a selection key as "OPCODE_NAME dest src1 src2 src3". Lives here
 // because the key layout belongs to this file's InstrKey.
-std::string FormatSequenceKey(uint32_t key);
+std::string FormatSequenceKey(uint64_t key);
 
 }  // namespace a64
 }  // namespace backend
